[K6.0] Remove namespaces on integration plugins because can't used
