+++ src/plugins/plg_kunena_community/KunenaLoginCommunity.php
@@ -10,12 +10,9 @@
  * @link            https://www.kunena.org
  **/
 
-namespace Kunena\Forum\Plugin\Kunena\Community;
-
 defined('_JEXEC') or die();
 
 use Joomla\CMS\Component\ComponentHelper;
-use function defined;
 
 /**
  * Class KunenaLoginCommunity
