+++ src/plugins/plg_kunena_community/KunenaPrivateCommunity.php
@@ -10,15 +10,11 @@
  * @link            https://www.kunena.org
  **/
 
-namespace Kunena\Forum\Plugin\Kunena\Community;
-
 defined('_JEXEC') or die();
 
-use Exception;
 use Joomla\CMS\Factory;
 use Joomla\CMS\Language\Text;
 use Kunena\Forum\Libraries\Integration\KunenaPrivate;
-use function defined;
 
 /**
  * Class KunenaPrivateCommunity
