+++ src/plugins/plg_kunena_easyblog/KunenaAvatarEasyblog.php
@@ -12,6 +12,9 @@
 defined('_JEXEC') or die();
 
 use Joomla\CMS\Uri\Uri;
+use Kunena\Forum\Libraries\Factory\KunenaFactory;
+use Kunena\Forum\Libraries\Integration\KunenaAvatar;
+use Kunena\Forum\Libraries\Route\KunenaRoute;
 
 /**
  * Class KunenaAvatarEasyblog
