+++ src/plugins/plg_kunena_easysocial/KunenaLoginEasySocial.php
@@ -15,12 +15,9 @@
  * See COPYRIGHT.php for copyright notices and details.
  */
 
-namespace Kunena\Forum\Plugin\Kunena\Easysocial;
-
 defined('_JEXEC') or die('Unauthorized Access');
 
 use Joomla\CMS\Component\ComponentHelper;
-use function defined;
 
 /**
  * @package     Kunena
