+++ src/plugins/plg_kunena_gravatar/KunenaAvatarGravatar.php
